Drop Default requirement thanks to generic default generation
The default values required by genericEditor are not used for display,
rather they are used to give a valid value to the editors for the non
selected constructors. This is an implementation detail of editorSum,
needed because threepenny doesn't support dynamic switching, and it is
better served by a generically generated value.

The change of Stackage snapshot is needed to pick up a new version
of constraints-sop with the HEAD type family, although we could as well
have defined it locally.