Improve the story around editing collections and add more examples
This change introduces a new member listEditor in the Editable class,
as well as new editor constructors editorList and editorCollection.

Moreover, the ListField type family has been added to cope with this change
in dual purpose datatypes.

Finally, the NonEmpty class has been renamed to HasEmpty and is now exported
on its own and a required superclass of Editable.